refactor: QuizTableをコンポーネント化
