+++ frontend/src/features/quiz/index.ts
@@ -1,4 +1,5 @@
 export * from './QuizForm'
 export * from './QuizFormModalButton'
+export * from './QuizTable'
 export * from './useQuiz'
 export * from './types'
